ProjectCenter fixes for compilation on 64-bit hosts.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/projectcenter/trunk@35330 72102866-910b-0410-8b05-ffd578937521

+++ Framework/PCProjectBuilder.m
@@ -870,8 +870,8 @@
 // --- Parsing utilities
 - (BOOL)line:(NSString *)lineString startsWithString:(NSString *)substring
 {
-  int      position = 0;
-  NSRange  range = NSMakeRange(position,1);
+  NSInteger position = 0;
+  NSRange   range = NSMakeRange(position,1);
 
   while ([[lineString substringFromRange:range] isEqualToString:@" "])
     {
@@ -1127,8 +1127,8 @@
     }
   else if ([components count] > 3)
     {
-      unsigned typeIndex;
-      NSString *substr;
+      NSUInteger typeIndex;
+      NSString   *substr;
 
       // file and includedFile
       file = [currentBuildPath
